+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Added a new cmdlet, `Find-SecurityTxtFile`, to locate a website's "security.txt" file, to ensure that it is in the correct well-known location instead of the legacy one.
 -   Changed how this module interfaces with `gpg`, so that it's more predictable on non-English systems.  Thank you to [Johannes Schöpp](https://github.com/jschpp) for [his great bug report, #3, and pull request, #4](https://github.com/rhymeswithmogul/SecurityTxtToolkit/pull/4).
 -   Improved detection of bad signatures.
+-   Remove extra spaces when a "security.txt" file has multiple `Preferred-Languages`.
 -   Include RFC 9116 in here as `about_SecurityTxt`.
 -   Fixed a bug where `Get-SecurityTxtFile` might attempt to fetch a file from a null URL.
 -   Code cleanup.